+++ apps/site-fake-deposit/src/translations/index.ts
@@ -1,0 +1,7 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+import { formatTranslations } from "@thecointech/shared/containers/LanguageProvider/initialize"
+
+export const translations = formatTranslations({
+  en: require('./en.json'),
+  fr: require('./fr.json')
+});