+++ libs/shared/src/containers/AccountMap/reducer.ts
@@ -4,11 +4,12 @@ import { AccountState, buildNewAccount } from "@thecointech/account";
 import { deleteAccount, storeAccount } from "@thecointech/account/store";
 import { IsValidAddress, NormalizeAddress } from "@thecointech/utilities";
 import { isLocal } from "@thecointech/signers";
-import { Wallet, Signer } from 'ethers';
 import { BaseReducer } from "../../store/immerReducer";
 import { mappedReducer, splitAccountFromRest } from "./mappedReducer";
 import { useSelector } from 'react-redux';
 import { activeAccountSelector, selectAccountArray } from './selectors';
+import { log } from '@thecointech/logging';
+import type { Signer } from '@ethersproject/abstract-signer';
 
 export class AccountMap extends BaseReducer<IAccountStoreAPI, AccountMapState>('accounts', initialState) implements IAccountStoreAPI {
 
@@ -24,41 +25,31 @@ export class AccountMap extends BaseReducer<IAccountStoreAPI, AccountMapState>('
       if (this.state.map[address] === undefined)
         throw new Error(`Cannot set activeAccount: ${address} does not exist`);
     }
-
+    log.trace({address}, 'Set active to {address}');
     this.draftState.active = address
   }
 
-  // Add a new account, optionally store in LocalStorate, in unlocked state
-  addAccount(name: string, address: string, signer: Signer, store: boolean = true, setActive: boolean = true, unlocked?: Wallet) {
-    // Check the signer & unlocked are actually the same account
-    if (unlocked && NormalizeAddress(address) != NormalizeAddress(unlocked?.address)) {
-      throw new Error("Accounts being stored are mis-matched");
-    }
+  // Add a new account
+  addAccount(name: string, address: string, signer: Signer) {
     // Create the account with default values
     const newAccount = buildNewAccount(name, address, signer);
-    // If asked to store, push to storage
-    if (store) {
-      storeAccount(newAccount);
-    }
-
-    // If we have an unlocked account, replace the signer
-    // (after the account has been pushed to storage)
-    if (unlocked) {
-      newAccount.signer = unlocked;
-    }
+    log.trace({address: newAccount.address}, 'Adding account: {address}');
+    // push to storage
+    storeAccount(newAccount);
     // Now for live storage
     this.draftState = {
       map: {
         ...this.state.map,
         [newAccount.address]: newAccount,
       } as any, // Requires use of any as the draftState automatically marks all members as mutable
-      active: setActive ? newAccount.address : this.state.active,
+      active: newAccount.address,
     };
   }
 
   // Remove the given account from list & storage
   deleteAccount(account: AccountState): void {
     const address = NormalizeAddress(account.address);
+    log.trace({address}, 'Removing account from localStorage: {address}');
     const { signer } = account;
     // We don't want active account with an invalid value
     if (this.state.active == address) {