+++ CHANGELOG.md
@@ -3,6 +3,97 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.40](https://github.com/thecointech/thecoin/compare/v0.2.39...v0.2.40) (2021-09-20)
+
+
+
+## [0.2.38](https://github.com/thecointech/thecoin/compare/v0.2.37...v0.2.38) (2021-09-18)
+
+
+
+## [0.2.37](https://github.com/thecointech/thecoin/compare/v0.2.36...v0.2.37) (2021-09-17)
+
+**Note:** Version bump only for package the-coin
+
+
+
+
+
+## [0.2.39](https://github.com/thecointech/thecoin/compare/v0.2.36...v0.2.39) (2021-09-20)
+
+**Note:** Version bump only for package the-coin
+
+
+
+
+
+## [0.2.37](https://github.com/thecointech/thecoin/compare/v0.2.36...v0.2.37) (2021-09-20)
+
+**Note:** Version bump only for package the-coin
+
+
+
+
+
+## [0.2.36](https://github.com/thecointech/thecoin/compare/v0.2.35...v0.2.36) (2021-09-17)
+
+**Note:** Version bump only for package the-coin
+
+
+
+
+
+## [0.2.35](https://github.com/thecointech/thecoin/compare/v0.2.34...v0.2.35) (2021-09-17)
+
+**Note:** Version bump only for package the-coin
+
+
+
+
+
+## [0.2.34](https://github.com/thecointech/thecoin/compare/v0.2.33...v0.2.34) (2021-09-17)
+
+**Note:** Version bump only for package the-coin
+
+
+
+
+
+## [0.2.33](https://github.com/thecointech/thecoin/compare/v0.2.32...v0.2.33) (2021-09-16)
+
+**Note:** Version bump only for package the-coin
+
+
+
+
+
+## [0.2.32](https://github.com/thecointech/thecoin/compare/v0.2.31...v0.2.32) (2021-09-16)
+
+**Note:** Version bump only for package the-coin
+
+
+
+
+
+## [0.2.31](https://github.com/thecointech/thecoin/compare/v0.2.30...v0.2.31) (2021-09-15)
+
+**Note:** Version bump only for package the-coin
+
+
+
+
+
+## [0.2.30](https://github.com/thecointech/thecoin/compare/v0.2.29...v0.2.30) (2021-09-15)
+
+
+### Bug Fixes
+
+* package.json to reduce vulnerabilities ([e3d66e9](https://github.com/thecointech/thecoin/commit/e3d66e936e443bbdeb357390fc64015962d13830))
+
+
+
+
+
 ## [0.2.29](https://github.com/thecointech/thecoin/compare/v0.2.28...v0.2.29) (2021-09-10)
 
 **Note:** Version bump only for package the-coin