+++ libs/signers/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.40](https://github.com/thecointech/thecoin/compare/v0.2.39...v0.2.40) (2021-09-20)
+
+**Note:** Version bump only for package @thecointech/signers
+
+
+
+
+
+## [0.2.39](https://github.com/thecointech/thecoin/compare/v0.2.36...v0.2.39) (2021-09-20)
+
+**Note:** Version bump only for package @thecointech/signers
+
+
+
+
+
+## [0.2.37](https://github.com/thecointech/thecoin/compare/v0.2.36...v0.2.37) (2021-09-20)
+
+**Note:** Version bump only for package @thecointech/signers
+
+
+
+
+
+## [0.2.34](https://github.com/thecointech/thecoin/compare/v0.2.33...v0.2.34) (2021-09-17)
+
+**Note:** Version bump only for package @thecointech/signers
+
+
+
+
+
+## [0.2.30](https://github.com/thecointech/thecoin/compare/v0.2.29...v0.2.30) (2021-09-15)
+
+**Note:** Version bump only for package @thecointech/signers
+
+
+
+
+
 ## [0.2.29](https://github.com/thecointech/thecoin/compare/v0.2.28...v0.2.29) (2021-09-10)
 
 **Note:** Version bump only for package @thecointech/signers